pythonBot/requirements.txt
asyncpg~=0.30.0
python-dotenv~=1.1.0
aiogram~=3.20.0.post0
aiosqlite~=0.21.0
pandas~=2.3.0